small fixes

+++ tests/test_common.h
@@ -6,7 +6,7 @@
 
 static void handle_assert(int passed, const char* assert, const char* file, long line) {
     if (!passed) {
-        printf("Assertion failed %s in %s at line %d\n", assert, file, line);
+        printf("Assertion failed %s in %s at line %i\n", assert, file, (int)line);
         exit(-1);
     }
 }
